refactored based on sonarcloud feedback
