--- java-itu-minitwit/src/main/java/RoP/Failure.java
@@ -1,38 +1,0 @@
-package RoP;
-
-public class Failure<T> implements Result<T> {
-
-    final Exception exception;
-
-    public Failure(Exception e) {
-        exception = e;
-    }
-
-    public Failure(String errorMsg) {
-        exception = new Exception(errorMsg);
-    }
-
-    @Override
-    public T get() {
-        throw new IllegalStateException(exception.getMessage());
-    }
-
-    @Override
-    public String getFailureMessage() {
-        return exception.getMessage();
-    }
-
-    @Override
-    public boolean isSuccess() {
-        return false;
-    }
-
-    public Exception getException() {
-        return exception;
-    }
-
-    @Override
-    public String toString() {
-        return exception.toString();
-    }
-}
