--- java-itu-minitwit/src/main/java/RoP/Result.java
@@ -1,9 +1,0 @@
-package RoP;
-
-public interface Result<T>
-{
-    public abstract T get();
-    public abstract String getFailureMessage();
-
-    public abstract boolean isSuccess();
-}
