+++ java-itu-minitwit/src/main/java/errorhandling/Result.java
@@ -1,0 +1,9 @@
+package errorhandling;
+
+public interface Result<T>
+{
+    public abstract T get();
+    public abstract String getFailureMessage();
+
+    public abstract boolean isSuccess();
+}
