+++ java-itu-minitwit/src/main/java/errorhandling/Success.java
@@ -1,0 +1,24 @@
+package errorhandling;
+
+public class Success<T> implements Result<T> {
+    private final T value;
+
+    public Success(T val) {
+        value = val;
+    }
+
+    @Override
+    public T get() {
+        return value;
+    }
+
+    @Override
+    public String getFailureMessage() {
+        throw new IllegalStateException("Type is success with "+value+" not an error");
+    }
+
+    @Override
+    public boolean isSuccess() {
+        return true;
+    }
+}
