+++ java-itu-minitwit/src/main/java/utilities/Statics.java
@@ -1,0 +1,9 @@
+package utilities;
+
+public class Statics {
+
+    private Statics() {}
+
+    public static int latest = 147371;
+
+}
