+++ java-itu-minitwit/src/test/java/RoPTests.java
@@ -1,6 +1,6 @@
-import RoP.Failure;
-import RoP.Result;
-import RoP.Success;
+import errorhandling.Failure;
+import errorhandling.Result;
+import errorhandling.Success;
 import org.junit.jupiter.api.Test;
 
 class RoPTests {
