--- java-itu-minitwit/src/test/java/UtilitiesTests.java
@@ -1,2 +1,0 @@
-public class UtilitiesTests {
-}
